fix linter issue
