--- src/dom.js
@@ -1,4 +1,0 @@
-const modal = document.querySelector('.popup-container');
-const comment = document.querySelectorAll('.comment');
-const Cards = document.querySelector('.cards');
-export { modal, comment, Cards };
