+++ src/index.html
@@ -1,51 +1,56 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script
-      src="https://kit.fontawesome.com/a11f8b435e.js"
-      crossorigin="anonymous"
-    ></script>
-    <title>Restaurant App</title>
-  </head>
 
-  <body>
-    <div class="popup-container hide"></div>
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <script src="https://kit.fontawesome.com/a11f8b435e.js" crossorigin="anonymous"></script>
+    <title>Restaurant App</title>
+</head>
+
+<body>
+    <div class="popup-container hide">
+        
+    </div>
     <header>
-      <nav class="nav">
-        <ul class="nav-items">
-          <li class="logo">
-            <i class="fa-solid fa-utensils fa-2x"></i>
-            <h1 class="logo-title">DeliciousFoody</h1>
-          </li>
-          <li class="nav-link">Categories(<b class="product-counter">1</b>)</li>
-        </ul>
-      </nav>
+        <nav class="nav">
+            <ul class="nav-items">
+                <li class="logo"><i class="fa-solid fa-utensils fa-2x"></i>
+                    <h1 class="logo-title">DeliciousFoody </h1>
+                </li>
+                <li class="nav-link">Categories(<b class="product-counter">1</b>)</li>
+            </ul>
+        </nav>
     </header>
     <main>
-      <div class="cards"></div>
-      <button id="4" class="comment">test</button>
+        <div class="cards">
+            
+         
+
+
+            </div>
+
     </main>
     <footer>
-      <div class="footer">
-        <div class="social-links">
-          <a href="">
-            <i class="fa fa-facebook fa-2x icon" aria-hidden="true"></i>
-          </a>
-          <a href="">
-            <i class="fa fa-twitter fa-2x icon" aria-hidden="true"></i>
-          </a>
-          <a href="">
-            <i class="fa fa-github fa-2x icon" aria-hidden="true"></i>
-          </a>
-          <a href="">
-            <i class="fa fa-envelope fa-2x icon" aria-hidden="true"></i>
-          </a>
+        <div class="footer">
+            <div class="social-links">
+                <a href="">
+                    <i class="fa fa-facebook fa-2x icon" aria-hidden="true"></i>
+                </a>
+                <a href="">
+                    <i class="fa fa-twitter fa-2x icon" aria-hidden="true"></i>
+                </a>
+                <a href="">
+                    <i class="fa fa-github fa-2x icon" aria-hidden="true"></i>
+                </a>
+                <a href="">
+                    <i class="fa fa-envelope fa-2x icon" aria-hidden="true"></i>
+                </a>
+            </div>
+            <p class="licence">&copy;Copyrights.All rights reserved</p>
         </div>
-        <p class="licence">&copy;Copyrights.All rights reserved</p>
-      </div>
     </footer>
-  </body>
+</body>
+
 </html>
