+++ src/modules/Api.js
@@ -1,4 +1,4 @@
-import DisplayCards from './Functions.js';
+import { DisplayPopup, DisplayCards } from './Functions.js';
 
 export default class Api {
   constructor() {
@@ -7,9 +7,27 @@ export default class Api {
     this.FreeMealEP = 'https://www.themealdb.com/api/json/v1/';
   }
 
-      GetMeals = async () => {
-        await fetch(`${this.FreeMealEP}/1/categories.php`)
-          .then((response) => response.json())
-          .then((json) => { DisplayCards(json.categories); });
-      }
+  GetMealInfos = async (id) => {
+    await fetch(`${this.FreeMealEP}/1/categories.php`)
+      .then((response) => response.json())
+      .then((json) => {
+        json.categories.forEach((item) => {
+          if (item.idCategory === id) {
+            this.GetExamples(item, id);
+          }
+        });
+      });
+  };
+
+  GetExamples = async (item, id) => {
+    await fetch(`${this.FreeMealEP}/1/filter.php?c=${item.strCategory}`)
+      .then((response) => response.json())
+      .then((json) => DisplayPopup(item, id, json));
+  };
+
+  GetMeals = async () => {
+    await fetch(`${this.FreeMealEP}/1/categories.php`)
+      .then((response) => response.json())
+      .then((json) => { DisplayCards(json.categories); });
+  }
 }
