+++ src/reset.css
@@ -120,14 +120,14 @@ blockquote:before,
 blockquote:after,
 q:before,
 q:after {
-  content: "";
+  content: '';
   content: none;
 }
 table {
   border-collapse: collapse;
   border-spacing: 0;
 }
-@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap");
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap');
 *,
 *::before,
 *::after {
@@ -141,7 +141,7 @@ table {
 
 body {
   background-color: #f2d2bd;
-  font-family: "Roboto", sans-serif;
+  font-family: 'Roboto', sans-serif;
   font-weight: bolder;
   letter-spacing: 0.2em;
   color: #6c757d;
@@ -170,6 +170,7 @@ body {
 .logo-title {
   font-size: 20px;
 }
+
 
 .form {
   display: flex;
@@ -221,14 +222,16 @@ body {
   font-weight: bolder;
 }
 
+
 .cards {
   display: flex;
   flex-wrap: wrap;
   justify-content: center;
-  align-items: center;
-  gap: 30px;
-  margin: 200px;
+  align-items: center; 
+gap: 30px;
+margin: 200px;
 }
+
 
 .social-links {
   display: flex;
@@ -241,6 +244,7 @@ body {
   margin-left: 2rem;
 }
 .footer {
+
   bottom: 1rem;
   width: 100%;
   margin: 1rem auto;
@@ -300,11 +304,13 @@ body {
   box-shadow: inset 0px 0px 10px 20px white;
 }
 
+
 .card {
   display: inline-block;
   width: 240px;
   text-align: center;
 }
+
 
 .img-popcontainer {
   height: 230px;
@@ -383,42 +389,43 @@ b {
   box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.2);
 }
 .img-container {
-  height: 230px;
-  width: 200px;
-  overflow: hidden;
-  border-radius: 0px 0px 20px 20px;
-  display: inline-block;
+    height: 230px;
+    width: 200px;
+    overflow: hidden;
+    border-radius: 0px 0px 20px 20px;
+    display: inline-block;
 }
 
 .img-container img {
-  transform: skew(0deg, -13deg);
-  height: 250px;
-  margin: -35px 0px 0px -70px;
+    transform: skew(0deg, -13deg);
+    height: 250px;
+    margin: -35px 0px 0px -70px;
 }
 
 .img {
-  display: inline-block;
-  border-radius: 20px;
-  overflow: hidden;
-  padding: 0px;
-  transform: skew(0deg, 13deg);
-  font-size: 0px;
-  margin: 30px 0px 0px 0px;
-  background: #c8c2c2;
-  height: 250px;
-  width: 200px;
+    display: inline-block;
+    border-radius: 20px;
+    overflow: hidden;
+    padding: 0px;
+    transform: skew(0deg, 13deg);
+    font-size: 0px;
+    margin: 30px 0px 0px 0px;
+    background: #c8c2c2;
+    height: 250px;
+    width: 200px;
 }
 
 .img-container img {
-  transition: transform 0.5s ease;
+  transition: transform .5s ease;
 }
 
 .img-container img:hover {
+
   transform: scale(0.8);
 }
 
 .title-container {
-  box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.2);
+  box-shadow: 0px 0px 10px 0px rgba(0,0,0,0.2);
 
   padding: 120px 20px 20px 20px;
   border-radius: 20px;
@@ -446,7 +453,7 @@ b {
 
 .comment {
   padding: 10px;
-  margin: 20px;
+  margin :20px;
   background-color: #04bcff;
   border: 0;
   border-radius: 5%;
